backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pymysql
httpx
python-dotenv
